refactor(heap-snapshot): derive the predecessor lists in one place
The V8 parser scattered the predecessor CSR lists itself, so every parser that
builds successor lists first would repeat that. `nodeAdjacencyGraphFromSuccessors`
now scatters them, and reports the caller's successor order unchanged.
